backend/admin/requirements.txt
# framework
Django>=4.2.4,<4.3.0

# other
psycopg[binary]==3.1.10
bcrypt>=4.0.1
django-cors-headers>=4.2.0
pydantic-settings>=2.0.2,<2.1.0
Pillow>=10.0.0,<10.1.0

# security
django-axes[ipware]>=6.1.1,<6.2.0

# model field
django-mdeditor==0.1.20
django-cleanup>=8.0.0,<8.1.0
django-colorfield>=0.10.0,<0.11.0

# static server
whitenoise>=6.5.0,<6.6.0
# asgi (wsgi) server
uvicorn>=0.23.2,<0.24.0
